requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.39.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asyncer==0.0.7
attrs==24.2.0
azure-core==1.32.0
azure-identity==1.19.0
beautifulsoup4==4.12.3
bidict==0.23.1
blis
boto3==1.35.58
botocore==1.35.58
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chainlit
numpy
charset-normalizer==3.4.0
chevron==0.14.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
fastai==2.7.18
fastapi==0.115.4
fastcore==1.7.19
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.16.1
filetype==1.2.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
google-auth==2.36.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
Lazify==0.4.0
literalai==0.0.623
llama-cloud
llama-index
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-ollama
llama-index-embeddings-openai
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-anthropic
llama-index-llms-azure-openai
llama-index-llms-ollama
llama-index-llms-openai
llama-index-multi-modal-llms-anthropic
llama-index-multi-modal-llms-azure-openai
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-parse
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
ollama==0.3.3
omegaconf==2.3.0
openai==1.54.3
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-exporter-otlp-proto-http==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
packaging==23.2
pandas==2.2.3
pillow==11.0.0
portalocker==2.10.1
preshed==3.0.9
propcache==0.2.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.9
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.10.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.4.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
starlette==0.41.2
striprtf==0.0.26
sympy==1.13.1
syncer==2.0.3
tenacity==8.5.0
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tomli==2.0.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uptrace==1.27.0
urllib3==2.2.3
uvicorn==0.25.0
wasabi==1.1.3
watchfiles==0.20.0
weasel==0.4.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.17.1
zipp==3.21.0


google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
pytube==15.0.0